Use es5-ext/global as a more robust way to resolve browser's window (#362)
The [es5-ext/global][es5g] module is based on Mathias Bynens' [polyfill][poly]
for [`globalThis`][gT]. It uses the same robust technique but is more
conservative in that it does not add a `globalThis` property to the resolved
object, i.e. `window` in a browser and `global` in Node. It also lacks the "old
IE" fallback seen in a [demo][demo] of the original polyfill, but that is
easily provided in this commit.

[es5g]: https://github.com/medikoo/es5-ext/blob/master/global.js
[poly]: https://mathiasbynens.be/notes/globalthis
[gT]: https://github.com/tc39/proposal-global
[demo]: https://mathiasbynens.be/demo/globalthis-ie